requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
jinja2
